DiffGitCached working for Windows
It was using a call to system() to call git and redirect to a temporary
file. But system already redirects to a different temporary file, so
the preview window was always empty. Rather clumsily, I now call system
to get the diff and write it to a new temporary file and display that.
Alternatively, I could call ":silent !git..." but that pops up a dos
box while git is running, unlike system